requirements.txt
fastapi
uvicorn[standard]
sqlmodel
pymysql
passlib[bcrypt]
pyjwt
python-jose[cryptography]
python-multipart
bcrypt>=4.0.1
pydantic-settings
psycopg2-binary
email-validator